add invariants in Profile

+++ backend/src/main/java/com/fasttasker/fast_tasker/application/dto/tasker/MinimalProfileResponse.java
@@ -12,7 +12,7 @@ public record MinimalProfileResponse(
         String firstName,
         String lastName,
         String photo,
-        int reputation,
+        float reputation,
         int clientReviews,
         int completedTasks
 ) {}
