+++ backend/src/main/java/com/fasttasker/fast_tasker/application/dto/tasker/ProfileRequest.java
@@ -8,7 +8,7 @@ public record ProfileRequest(
         String lastName,
         String photo,
         String about,
-        int reputation,
+        float reputation,
         int clientReviews,
         int completedTasks,
         LocationRequest location
